add app module to libcommon, move common headers
The app module handles the menial tasks of setting up the INPUT_FILE
and OUTPUT_FILE positional arguments, opening the corresponding files,
and ensuring that the files and utility-specific state is cleaned up
no matter what errors occur. It is generic enough that the zlib_common
module is no longer necessary, except perhaps to convert from
AppIOState information to pointers and lengths to a z_stream.

Moving the common headers is mostly a convenience factor so that headers
of my creation are easily disambugated from system or library headers.