bump version to 0.2.2, add common/mmc.h
common/mmc.h contains the MMC_VERSION and MMC_AUTHOR preprocessor
symbols, which are to simplify my job when it comes to incrementing
the version number. Perhaps in the future the file mmc.h will actually
be a file mmc.h.in that is then processed by CMake to create mmc.h.